+++ qcop/adapters/terachem_pbs.py
@@ -1,0 +1,13 @@
+from .terachem_fe import TeraChemFEAdapter
+
+
+class TeraChemPBSAdapter(TeraChemFEAdapter):
+    """Adapter for TeraChem's Protocol Buffer Server."""
+
+    program = "terachem-pbs"
+
+    def __init__(self):
+        super().__init__()
+        # Check that xtb-python is installed.
+        self.tcpb = self._ensure_tcpb()
+        self.client = self.tcpb.TCProtobufClient